+++ citizen-intelligence-agency/pom.xml
@@ -1694,15 +1694,8 @@
           <excludes>
             <exclude>**/*$*</exclude>
           </excludes>
-          <systemPropertyVariables>
-            <java.util.logging.config.file>src/test/resources/logging.properties</java.util.logging.config.file>
-          </systemPropertyVariables>
-          <argLine>
-            --add-opens java.base/java.lang=ALL-UNNAMED
-            --add-opens java.base/java.util=ALL-UNNAMED
-            --add-opens java.base/java.lang.reflect=ALL-UNNAMED
-            -Xmx2048m
-          </argLine>
+          <argLine>${surefireArgLine} ${testAdditionalRuntimeArgLine}</argLine>
+          <useModulePath>false</useModulePath>
           <forkCount>1</forkCount>
           <reuseForks>false</reuseForks>
         </configuration>
