+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/AbstractUITest.java
@@ -1,11 +1,8 @@
 package com.hack23.cia.systemintegrationtest;
 
 import org.junit.After;
-import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.openqa.selenium.WebDriver;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import com.hack23.cia.systemintegrationtest.ui.TestConstants;
 import com.hack23.cia.systemintegrationtest.ui.UserPageVisit;
@@ -16,16 +13,11 @@ import com.hack23.cia.systemintegrationtest.ui.WebDriverFactory;
  */
 public abstract class AbstractUITest extends AbstractRoleSystemITest {
 
-	/** The Constant LOG. */
-	private static final Logger LOG = LoggerFactory.getLogger(AbstractUITest.class);
 
 	/** The Constant DEFAULT_TIMEOUT. */
 	protected static final long DEFAULT_TIMEOUT = TestConstants.DEFAULT_TIMEOUT;
 
-	/** The Constant BASE_URL. */
-	protected static final String BASE_URL = System.getProperty("system.test.target.url",
-			CitizenIntelligenceAgencyServer.ACCESS_URL);
-
+	
 	/** The driver. */
 	protected static WebDriver driver;
 
@@ -37,21 +29,10 @@ public abstract class AbstractUITest extends AbstractRoleSystemITest {
 	 */
 	@BeforeClass
 	public static void globalSetup() {
-		LOG.info("Setting up test with browser");
 		driver = WebDriverFactory.createDriver();
 		pageVisit = new UserPageVisit(driver);
 	}
 
-	/**
-	 * Global teardown.
-	 */
-	@AfterClass
-	public static void globalTeardown() {
-		LOG.info("Tearing down WebDriver after all tests have run");
-		if (driver != null) {
-			driver.quit();
-		}
-	}
 
 	@After
 	public void cleanBrowser() {
