+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/suites/UnitTestSuite.java
@@ -48,5 +48,5 @@ import com.hack23.cia.web.impl.ui.application.views.pageclicklistener.UpdateSear
     ListPropertyConverterTest.class,
     UserContextUtilTest.class})
 public class UnitTestSuite {
-		
+
 }
